Refactor app architecture
The app is now structured into 2 layer: data and ui.

Every composable screen accepts a state, the state, in those cases, is derived from the main view model state.

The data layer is organized into data sources and repositories

The ViewModel shouldn't instantiate the repositories, I'll fix it in a future commit